refactor: move unit tests to own files
